Add Backend components module and converter functionality
This commit adds a new module called "components" and includes the implementation of the "converter" functionality. The "components" module contains two sub-modules: "vector" and "converter". The "converter" module includes a function called "convert" that is yet to be implemented. Additionally, a new file called "Cargo.toml" is added to the project, which includes dependencies for serde, rocket, and rocket_cors. Finally, a new file called "things.txt" is added to the "components" directory, which contains some placeholder text.